+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Changed
+- In case `flye` fails due to low coverage, the workflow will continue and this will be indicated in the report.
+
 ## [v0.2.14]
 ### Added
 - `--isolates` parameter that will run the ResFinder tool on the final assembly to output antimicrobial resistance genes.